Added logging after successful matches.
Previously it was a bit of black box as far as debugging and understanding
the algorithm went. There are now logging calls throughout the code which
use a stub .log() method call. This should be implemented by the calling
code to get logging output (using print() or console.log() probably,
depending on your environment).