Measure checks now handle words starting with Y.
The test cases "ycleped" and "ycliped" were failing because the regexs
for the various measure checks treated a leading Y as a vowel. The
algorithm paper isn't specific about a leading Y being a vowel or consonant
but I think the test case clears that up.

Sadly the regexs are now hideously long and ugly and could probably be
written in a far nicer way.